moves serialize function
