+++ src/lambda/serializeJs.ts
@@ -1,10 +1,10 @@
 import * as t from '@babel/types';
-import {Term, TermType} from '../term';
+import {Term, TermType} from './term';
 
 /**
  * Compiles a lambda calculus term to a JavaScript expression.
  */
-export function serialize(term: Term): t.Expression {
+export function serializeJs(term: Term): t.Expression {
   switch (term.type) {
     case TermType.Variable: {
       return t.identifier(term.name);
@@ -12,12 +12,12 @@ export function serialize(term: Term): t.Expression {
     case TermType.Abstraction: {
       return t.arrowFunctionExpression(
         [t.identifier(term.parameter)],
-        serialize(term.body),
+        serializeJs(term.body),
       );
     }
     case TermType.Application: {
-      return t.callExpression(serialize(term.callee), [
-        serialize(term.argument),
+      return t.callExpression(serializeJs(term.callee), [
+        serializeJs(term.argument),
       ]);
     }
   }
